Extract collaborator queries to hooks
